models/StarknetModel.py
import numpy as np
from sklearn.ensemble import RandomForestClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.metrics import accuracy_score, classification_report, confusion_matrix

class StarknetModel:
    def __init__(self):
        self.model = None
        self.scaler = StandardScaler()
        self.imputer = SimpleImputer(strategy='mean')
        self.column_transformer = ColumnTransformer(
            transformers=[
                ('num', self.scaler, ['feature1', 'feature2', 'feature3']),
                ('cat', self.imputer, ['feature4', 'feature5'])
            ]
        )
        self.pipeline = Pipeline([
            ('column_transformer', self.column_transformer),
            ('random_forest', RandomForestClassifier(n_estimators=100))
        ])

    def preprocess(self, input_data):
        # Preprocess input data
        # Convert categorical variables to numerical variables
        # and scale/normalize the data
        input_data['feature4'] = input_data['feature4'].astype('category')
        input_data['feature5'] = input_data['feature5'].astype('category')
        input_data['feature4'] = input_data['feature4'].cat.codes
        input_data['feature5'] = input_data['feature5'].cat.codes
        return self.column_transformer.fit_transform(input_data)

    def predict(self, preprocessed_data):
        # Make predictions using the model
        return self.pipeline.predict(preprocessed_data)

    def postprocess(self, predictions):
        # Postprocess predictions
        # Convert numerical predictions to categorical labels
        return np.array(['class1' if x == 0 else 'class2' for x in predictions])

    def is_valid_input(self, input_data):
        # Check if the input data is valid
        # Check if the input data is of the correct shape and type
        if input_data.shape[1] != 5:
            return False
        if not all([x in input_data.columns for x in ['feature1', 'feature2', 'feature3', 'feature4', 'feature5']]):
            return False
        return True

    def is_valid_output(self, output_data):
        # Check if the output data is valid
        # Check if the output data is of the correct shape and type
        if output_data.shape[0] != self.pipeline.predict(self.column_transformer.fit_transform(self.input_data)).shape[0]:
            return False
        if not all([x in ['class1', 'class2'] for x in output_data]):
            return False
        return True

    def update(self, new_data):
        # Update the machine learning model
        # Retrain the model on the new data
        self.pipeline.fit(self.column_transformer.fit_transform(new_data), new_data['target'])

    def save(self):
        # Save the updated model
        # Save the model to a file
        import joblib
        joblib.dump(self.pipeline, 'starknet_model.joblib')

    def evaluate(self, input_data, target):
        # Evaluate the model
        # Calculate accuracy, classification report, and confusion matrix
        predictions = self.predict(self.preprocess(input_data))
        accuracy = accuracy_score(target, predictions)
        report = classification_report(target, predictions)
        matrix = confusion_matrix(target, predictions)
        return accuracy, report, matrix