requirements.txt
sklearn
numpy
starknet_py
giza-sdk
asyncio
